# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Klinik Service API",
        "description": "Klinik Service API",
        "contact": {
            "email": "acxell@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Primary API Server"
        }
    ],
    "paths": {
        "/appointment": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create an appointment",
                "description": "Create a new appointment",
                "operationId": "createAppointment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "diagnose_id"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "diagnose_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/appointment/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment details",
                "description": "Retrieve ID",
                "operationId": "getAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update appointment status",
                "description": "Update Appointment Status",
                "operationId": "updateAppointmentStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/diagnose": {
            "post": {
                "tags": [
                    "Diagnoses"
                ],
                "summary": "Create a new diagnose",
                "operationId": "fdc4a8bfa30720d09d0210bb9611170f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "service"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Flu"
                                    },
                                    "service": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Diagnose created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Diagnose"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/patient": {
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Create a new patient",
                "operationId": "6a25fc719b702841b2d92860ecd2f4b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/service": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "operationId": "8a208399289a320e49f0c2d6c5e5db0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "General Checkup"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Service"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AppointmentRequest": {
                "title": "Appointment Request",
                "properties": {
                    "patient_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "diagnose_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "integer",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "Appointment": {
                "title": "Appointment",
                "description": "Appointment model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "patient_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "diagnose_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-14T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Diagnose": {
                "title": "Diagnose",
                "description": "Diagnose model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Flu"
                    },
                    "service": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-14T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Patient": {
                "title": "Patient",
                "description": "Patient model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acxell"
                    }
                },
                "type": "object"
            },
            "Service": {
                "title": "Service",
                "description": "Service model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Layanan Periksa"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-14T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Diagnoses",
            "description": "Diagnose Management"
        },
        {
            "name": "Patients",
            "description": "Patient Management"
        },
        {
            "name": "Services",
            "description": "Service Management"
        },
        {
            "name": "Appointments",
            "description": "Appointments"
        }
    ]
}